requirements.txt
streamlit
streamlit-folium
folium
pandas
openrouteservice
python-dotenv
geopy
requests
openmeteo-requests
requests-cache
retry-requests